Refactor IPlayer to IActor, IPlayer, INonPlayer, add EasyAI
