+++ src/main/java/biblioteket/roborally/actors/INonPlayer.java
@@ -1,0 +1,7 @@
+package biblioteket.roborally.actors;
+
+import biblioteket.roborally.programcards.ICardDeck;
+
+public interface INonPlayer extends IActor {
+    void chooseCards(ICardDeck deck);
+}
