+++ src/main/java/biblioteket/roborally/elements/interacting/HoleElement.java
@@ -1,12 +1,12 @@
 package biblioteket.roborally.elements.interacting;
 
-import biblioteket.roborally.actors.IPlayer;
+import biblioteket.roborally.actors.IActor;
 import biblioteket.roborally.actors.IRobot;
 
 public class HoleElement implements InteractingElement {
 
     @Override
-    public void interact(IPlayer player) {
+    public void interact(IActor player) {
         IRobot robot = player.getRobot();
         player.removeOneLife();
         robot.moveToArchiveMarker();
