+++ src/main/java/biblioteket/roborally/elements/interacting/SingleWrenchRepairElement.java
@@ -1,6 +1,6 @@
 package biblioteket.roborally.elements.interacting;
 
-import biblioteket.roborally.actors.IPlayer;
+import biblioteket.roborally.actors.IActor;
 import biblioteket.roborally.actors.IRobot;
 
 public class SingleWrenchRepairElement implements InteractingElement {
@@ -11,7 +11,7 @@ public class SingleWrenchRepairElement implements InteractingElement {
      * @param player the player with the robot on current cell
      */
     @Override
-    public void interact(IPlayer player) {
+    public void interact(IActor player) {
         IRobot robot = player.getRobot();
         robot.removeDamageTokens(1);
         robot.setArchiveMarker(robot.getPosition());
