+++ src/main/java/biblioteket/roborally/elements/interacting/cogs/LeftRotatingCogElement.java
@@ -1,11 +1,11 @@
 package biblioteket.roborally.elements.interacting.cogs;
 
-import biblioteket.roborally.actors.IPlayer;
+import biblioteket.roborally.actors.IActor;
 
 public class LeftRotatingCogElement extends CogElement {
 
     @Override
-    public void interact(IPlayer player) {
+    public void interact(IActor player) {
         player.rotateRobot(false, 500);
     }
