+++ src/main/java/biblioteket/roborally/elements/interacting/cogs/RightRotatingCogElement.java
@@ -1,11 +1,11 @@
 package biblioteket.roborally.elements.interacting.cogs;
 
-import biblioteket.roborally.actors.IPlayer;
+import biblioteket.roborally.actors.IActor;
 
 public class RightRotatingCogElement extends CogElement {
 
     @Override
-    public void interact(IPlayer player) {
+    public void interact(IActor player) {
         player.rotateRobot(true, 500);
     }
