+++ src/main/java/biblioteket/roborally/elements/interacting/conveyorbelts/ConveyorBeltElement.java
@@ -1,6 +1,6 @@
 package biblioteket.roborally.elements.interacting.conveyorbelts;
 
-import biblioteket.roborally.actors.IPlayer;
+import biblioteket.roborally.actors.IActor;
 import biblioteket.roborally.board.Direction;
 import biblioteket.roborally.elements.interacting.InteractingElement;
 
@@ -17,7 +17,7 @@ public class ConveyorBeltElement implements InteractingElement {
      * @param player with robot to be moved
      */
     @Override
-    public void interact(IPlayer player) {
+    public void interact(IActor player) {
         player.moveRobot(direction, 500, false);
     }
