+++ src/test/java/biblioteket/roborally/elements/interacting/CogTest.java
@@ -1,8 +1,8 @@
 package biblioteket.roborally.elements.interacting;
 
 import biblioteket.roborally.TestRunner;
-import biblioteket.roborally.actors.IPlayer;
-import biblioteket.roborally.actors.Player;
+import biblioteket.roborally.actors.IActor;
+import biblioteket.roborally.actors.Actor;
 import biblioteket.roborally.actors.Robot;
 import biblioteket.roborally.actors.RobotRenderer;
 import biblioteket.roborally.board.Board;
@@ -23,7 +23,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class CogTest {
     private static IBoard board;
     private static RobotRenderer robotRenderer;
-    private IPlayer player;
+    private IActor player;
 
     @BeforeAll
     private static void setup() {
@@ -33,7 +33,7 @@ public class CogTest {
 
     @BeforeEach
     void setUp() {
-        player = new Player(board, null, null, robotRenderer);
+        player = new Actor(board, null, null, robotRenderer);
         player.setRobot(new Robot(new ArchiveMarkerElement(1)));
     }
